Conflicts:
	Engine/Source/Shared/System/Content/CsCore.cpp
	Engine/Source/Shared/System/Debug/DsTemplate.cpp
	Engine/Source/Shared/System/Renderer/D3D11/RsContextD3D11.cpp
	Engine/Source/Shared/System/Scene/Import/ScnAnimationImport.cpp
	Engine/Source/Shared/System/Scene/Import/ScnModelImport.cpp

+++ Engine/Source/Shared/System/Scene/Import/ScnAnimationImport.cpp
@@ -190,7 +190,7 @@ BcBool ScnAnimationImport::import( const Json::Value& )
 	if( Scene_ != nullptr )
 	{
 		PSY_LOG( "Found %u animations:\n", Scene_->mNumAnimations );
-		for( int Idx = 0; Idx < Scene_->mNumAnimations; ++Idx )
+		for( int Idx = 0; Idx < (int)Scene_->mNumAnimations; ++Idx )
 		{
 			PSY_LOG( " - %s\n", Scene_->mAnimations[ Idx ]->mName.C_Str() );
 		}
